+++ js/front/dataset/index.js
@@ -153,7 +153,7 @@ new Vue({
          * @param  {Object} resource A resource as extracted from JSON-LD
          */
         checkResource(resource) {
-            const url = parseUrl(resource.url);
+            const url = parseUrl(resource.contentUrl);
             const resource_el = document.querySelector(`#resource-${resource['@id']}`);
             const el = resource_el.querySelector('.format-label');
             const checkurl = resource_el.dataset.checkurl;